+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Add `EnvelopeProperties.declineAllowed` and `EnvelopeProperties.declineReasonRequired`
 - Add `EnvelopeRecipient.declineReason`
 - Add `EnvelopeRecipientAttachment` resource
+- Add `EnvelopeRecipientEndpoint.attachments`
 
 ## [1.5.0] - 2021-11-09
 ### Added